ventilate meta.py to meta/
