+++ fp/meta/__init__.py
@@ -1,0 +1,3 @@
+from .kind      import Kind
+from .type      import Type
+from .functor   import Functor, Bifunctor
